Use u32 for block heights
Since Tendermint uses i64 to represent heights the only sensible option
is to use a u32 in gears.

+++ gears/src/baseapp/mode/check.rs
@@ -44,7 +44,7 @@ impl<DB: Database, AH: ABCIHandler> ExecutionMode<DB, AH> for CheckTxMode<DB, AH
 
     fn build_ctx(
         &mut self,
-        height: u64,
+        height: u32,
         header: Header,
         consensus_params: ConsensusParams,
         fee: Option<&Fee>,
